usersnap #4137 swap "able bodied" with "non-disabled"
Sorry if this is a duplicate. I thought I put in a PR but then couldn't find any history of it. Student pointed out that "able bodied" is considered a microaggression. I followed this guide  https://ncdj.org/style-guide/ that is provided as a resource for reporting on disabled persons.